2.7.0 - The Game Status Update
* now only suggests channels that you can read and send to when you try
to run a command in the wrong channel.
+ GroupUp command
+ Groups Channel setting added
+ ClearGroupUp command
+ SetPlayingStatus command
* fixed a bug involving the generation of the Guild Files.
* started work on wiki page generation.
+ playing statuses that randomly select a game from chucklefish's
published games, starbound 2/5, Wargroove 1/5, all others 1/25
+ added website blacklist check to url variables.
+ QuickRespond DM command
+ started implementation of Sail's Pixel xp system, (is now tracking XP)
* Fixed a bug preventing messages from being deleted when adding the X
emoji as a reaction to a message.
+ added a system for users to add up to 5 links to their profile via the
EditLinks command.
+ XpDenied Channel Setting adde

+++ src/main/java/Commands/Creator/GetMessageData.java
@@ -20,11 +20,11 @@ public class GetMessageData implements Command {
         int totalAvg = 0;
         for (IGuild guild : command.client.getGuilds()) {
             String builder = "";
-            int totalUsers = 0;
-            int totalMessage = 0;
-            int totalMessageAvg = 0;
-            int topGuild = 0;
-            int bottomGuild = 0;
+            long totalUsers = 0;
+            long totalMessage = 0;
+            long totalMessageAvg = 0;
+            long topGuild = 0;
+            long bottomGuild = 0;
             String topUser = null;
             String bottomUser = null;
             GuildContentObject content = Globals.getGuildContent(guild.getStringID());
